Renamed "Common" files to "TIGRE_common"
