+++ Python/tigre/Source/POCS_TV2.hpp
@@ -52,7 +52,7 @@ Codes  : https://github.com/CERN/TIGRE
 
 #ifndef POCS_TV_HPP
 #define POCS_TV_HPP
-#include "Common.hpp"
+#include "TIGRE_common.hpp"
 void aw_pocs_tv(float* img,float* dst,float alpha,const long* image_size, int maxIter,const float delta);
 void checkFreeMemory(int deviceCount,size_t *mem_GPU_global);
