+++ Python/tigre/Source/Siddon_projection.cu
@@ -51,7 +51,7 @@
 #include <cuda_runtime_api.h>
 #include <cuda.h>
 #include "Siddon_projection.hpp"
-#include "Common.hpp"
+#include "TIGRE_common.hpp"
 #include <math.h>
 
 #define cudaCheckErrors(msg) \
