+++ Python/tigre/Source/ray_interpolated_projection_parallel.cu
@@ -53,7 +53,7 @@
 #include <cuda_runtime_api.h>
 #include <cuda.h>
 #include "ray_interpolated_projection_parallel.hpp"
-#include "Common.hpp"
+#include "TIGRE_common.hpp"
 #include <math.h>
 
 #define cudaCheckErrors(msg) \
