+++ Python/tigre/Source/voxel_backprojection.cu
@@ -48,7 +48,7 @@
 #include <cuda_runtime_api.h>
 #include <cuda.h>
 #include "voxel_backprojection.hpp"
-#include "Common.hpp"
+#include "TIGRE_common.hpp"
 #include <math.h>
 
 // https://stackoverflow.com/questions/16282136/is-there-a-cuda-equivalent-of-perror
